Ml-Model-Backend-main/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
tensorflow==2.20.0
numpy==2.2.6
Pillow==12.0.0
gunicorn==23.0.0